+++ tests/compile/r_pernici1.pyx
@@ -8,16 +8,16 @@ __doc__ = u"""
 """
 
 cdef extern from "math.h":
-  double M_PI
+  f64 M_PI
 
 #cdef unsigned long int n1
 #n1 = 4293858116
 
-cdef double pi
+cdef f64 pi
 pi = 3.14159265358979323846
 
 def main():
   #print n1
   print "%.18f" % M_PI
-  print "%.18f" % (<float> M_PI)
+  print "%.18f" % (<f64> M_PI)
   print "%.18f" % pi
