+++ tests/memoryview/cythonarray.pyx
@@ -9,7 +9,6 @@ cimport cython as cy
 
 include "../testsupport/cythonarrayutil.pxi"
 
-
 def length(shape):
     """
     >>> len(length((2,)))
@@ -19,10 +18,9 @@ def length(shape):
     >>> len(length((5,3,2)))
     5
     """
-    cdef array cvarray = array(shape=shape, itemsize=sizeof(int), format="i", mode='c')
+    cdef array cvarray = array(shape=shape, itemsize=sizeof(i32), format="i", mode='c')
     assert len(cvarray) == shape[0]
     return cvarray
-
 
 def contiguity():
     '''
@@ -35,18 +33,18 @@ def contiguity():
     2 3
     2
     '''
-    cdef v.array cvarray = cy.view.array(shape=(2,3), itemsize=sizeof(int), format="i", mode='c')
-    assert cvarray.len == 2*3*sizeof(int), (cvarray.len, 2*3*sizeof(int))
-    assert cvarray.itemsize == sizeof(int)
+    cdef v.array cvarray = cy.view.array(shape=(2, 3), itemsize=sizeof(i32), format="i", mode='c')
+    assert cvarray.len == 2*3*sizeof(i32), (cvarray.len, 2*3*sizeof(i32))
+    assert cvarray.itemsize == sizeof(i32)
     print cvarray.strides[0], cvarray.strides[1]
     print cvarray.shape[0], cvarray.shape[1]
     print cvarray.ndim
 
     print
 
-    cdef v.array farray = v.array(shape=(2,3), itemsize=sizeof(int), format="i", mode='fortran')
-    assert farray.len == 2*3*sizeof(int)
-    assert farray.itemsize == sizeof(int)
+    cdef v.array farray = v.array(shape=(2, 3), itemsize=sizeof(i32), format="i", mode='fortran')
+    assert farray.len == 2*3*sizeof(i32)
+    assert farray.itemsize == sizeof(i32)
     print farray.strides[0], farray.strides[1]
     print farray.shape[0], farray.shape[1]
     print farray.ndim
@@ -55,30 +53,30 @@ def acquire():
     '''
     >>> acquire()
     '''
-    cdef object[int, ndim=1, mode="c"] buf1d = \
-            array(shape=(10,), itemsize=sizeof(int), format='i', mode='c')
-    cdef object[int, ndim=2, mode="c"] buf2d = \
-            array(shape=(10,10), itemsize=sizeof(int), format='i')
-    cdef object[unsigned long, ndim=3, mode='fortran'] buf3d = \
-            array(shape=(1,2,3), itemsize=sizeof(unsigned long), format='L', mode='fortran')
+    cdef object[i32, ndim=1, mode="c"] buf1d = \
+            array(shape=(10,), itemsize=sizeof(i32), format='i', mode='c')
+    cdef object[i32, ndim=2, mode="c"] buf2d = \
+            array(shape=(10,10), itemsize=sizeof(i32), format='i')
+    cdef object[u64, ndim=3, mode='fortran'] buf3d = \
+            array(shape=(1, 2, 3), itemsize=sizeof(unsigned long), format='L', mode='fortran')
     cdef object[long double, ndim=3, mode='fortran'] bufld = \
-            array(shape=(1,2,3), itemsize=sizeof(long double), format='g', mode='fortran')
+            array(shape=(1, 2, 3), itemsize=sizeof(long double), format='g', mode='fortran')
 
 def full_or_strided():
     '''
     >>> full_or_strided()
     '''
-    cdef object[float, ndim=2, mode='full'] fullbuf = \
-            array(shape=(10,10), itemsize=sizeof(float), format='f', mode='c')
-    cdef object[long long int, ndim=3, mode='strided'] stridedbuf = \
-            array(shape=(1,2,3), itemsize=sizeof(long long int), format='q', mode='fortran')
+    cdef object[f32, ndim=2, mode='full'] fullbuf = \
+            array(shape=(10, 10), itemsize=sizeof(f32), format='f', mode='c')
+    cdef object[i128, ndim=3, mode='strided'] stridedbuf = \
+            array(shape=(1, 2, 3), itemsize=sizeof(i128), format='q', mode='fortran')
 
 def dont_allocate_buffer():
     """
     >>> dont_allocate_buffer()
     callback called
     """
-    cdef array result = array((10, 10), itemsize=sizeof(int), format='i', allocate_buffer=False)
+    cdef array result = array((10, 10), itemsize=sizeof(i32), format='i', allocate_buffer=False)
     assert result.data == NULL
     result.callback_free_data = callback
     result = None
@@ -91,8 +89,8 @@ def test_cython_array_getbuffer():
     98
     61
     """
-    cdef int[:, ::1] cslice = create_array((14, 10), 'c')
-    cdef int[::1, :] fslice = create_array((14, 10), 'fortran')
+    cdef i32[:, ::1] cslice = create_array((14, 10), 'c')
+    cdef i32[::1, :] fslice = create_array((14, 10), 'fortran')
 
     print cslice[9, 8]
     print cslice[6, 1]
@@ -117,10 +115,10 @@ def test_cython_array_index():
     print f_array[9, 8]
     print f_array[6, 1]
 
-cdef int *getp(int dim1=10, int dim2=10, dim3=1) except NULL:
+cdef i32 *getp(i32 dim1=10, i32 dim2=10, dim3=1) except NULL:
     print "getp()"
 
-    cdef int *p = <int *> malloc(dim1 * dim2 * dim3 * sizeof(int))
+    cdef i32 *p = <i32 *> malloc(dim1 * dim2 * dim3 * sizeof(i32))
 
     if p == NULL:
         raise MemoryError
@@ -151,28 +149,28 @@ def test_array_from_pointer():
     119
     callback free data called
     """
-    cdef int *p = getp()
-    cdef array c_arr = <int[:10, :10]> p
+    cdef i32 *p = getp()
+    cdef array c_arr = <i32[:10, :10]> p
     c_arr.callback_free_data = callback_free_data
     print c_arr[6, 9]
     print c_arr.mode
 
-    c_arr = (<int[:10:1, :10]> getp())
+    c_arr = (<i32[:10:1, :10]> getp())
     print c_arr.mode
     c_arr.callback_free_data = free
 
-    c_arr =  <int[:10, :10]> getp()
+    c_arr =  <i32[:10, :10]> getp()
     c_arr.callback_free_data = free
-    cdef int[:, ::1] mslice = c_arr
+    cdef i32[:, ::1] mslice = c_arr
     print mslice[5, 6]
 
-    c_arr = <int[:12, :10]> getp(12, 10)
+    c_arr = <i32[:12, :10]> getp(12, 10)
     c_arr.callback_free_data = free
     print c_arr[5, 6]
 
-    cdef int m = 12
-    cdef int n = 10
-    c_arr = <int[:m, :n]> getp(m, n)
+    cdef i32 m = 12
+    cdef i32 n = 10
+    c_arr = <i32[:m, :n]> getp(m, n)
     c_arr.callback_free_data = callback_free_data
     print c_arr[m - 1, n - 1]
 
@@ -183,12 +181,12 @@ def test_array_from_pointer_3d():
     3 3
     True True
     """
-    cdef int *p = getp(2, 2, 2)
-    cdef array c_arr = <int[:2, :2, :2:1]> p
-    cdef array f_arr = <int[:2:1, :2, :2]> p
+    cdef i32 *p = getp(2, 2, 2)
+    cdef array c_arr = <i32[:2, :2, :2:1]> p
+    cdef array f_arr = <i32[:2:1, :2, :2]> p
 
-    cdef int[:, :, ::1] m1 = c_arr
-    cdef int[::1, :, :] m2 = f_arr
+    cdef i32[:, :, ::1] m1 = c_arr
+    cdef i32[::1, :, :] m2 = f_arr
 
     print m1[0, 1, 1], m2[1, 1, 0]
     print m1.is_c_contig(), m2.is_f_contig()
@@ -199,12 +197,12 @@ def test_cyarray_from_carray():
     0 8 21
     0 8 21
     """
-    cdef int[7][8] a
+    cdef i32[7][8] a
     for i in range(7):
         for j in range(8):
             a[i][j] = i * 8 + j
 
-    cdef int[:, :] mslice = <int[:, :]> a
+    cdef i32[:, :] mslice = <i32[:, :]> a
     print mslice[0, 0], mslice[1, 0], mslice[2, 5]
 
     mslice = a
@@ -217,7 +215,6 @@ class InheritFrom(v.array):
     >>> inst = InheritFrom(shape=(3, 3, 3), itemsize=4, format="i")
     """
     pass
-
 
 def test_char_array_in_python_api(*shape):
     """
@@ -321,4 +318,3 @@ match arr:
         assert globs['res']
 
     return isinstance(arr, Sequence)
-
